Define `importLoaders` in `vue.config.js`
Because `vue-loader` and `vue-cli` have updated since the last build,
this commit updates the updates the `vue.config.js` and defines the
`importLoaders` on the `css-loader`. This allows the app to use `*.scss`
files, in conjunctor with the CSS Modules `composes` spec. Without
defining the `importLoaders`, when the `css-loader` encounters composed
`.scss` files, webpack passes raw SASS to the `postcss-loader`, which
throws an error.

See (github
issue)[https://github.com/webpack-contrib/css-loader/issues/137]