Fix formatting of Edit button
This commit normalizes the Edit button's font on the "Well done!" page.
The previous font-family was "lato-bold-normal", but needed to be
"lato-regular-normal" like the rest of the font in it's row.

To do this we add a `weight` prop to the `BaseButtonLink` component,
which is used to calculate the `weightClass` computed value. This is
similar to how the `size` and `role` props our handled.

To be able to change this particular button without affecting the other
buttons on the site, we make `bold` the default and manually add
regular weight as a prop to the edit button.