Upgrades to vue-selector 3.x
This commit fixes changes to `vue-selector` made when converting to 3.x.
We change vSelect specific classes to be prepended with `vs__`, making
them less likely to conflict with personal CSS classes. We also have to
import the css file from the module manually because some developer
thought it was a good idea to break everyones builds.

This commit also fixes minor `object-curly-spacing` errors generated by
eslint.