Closes flyout when body is clicked
This commit adds an event listener to the `body` that closes any open
`BaseFlyout` by dispatching an action to the store signaling that all
flyouts are closed. We do this in `ActivitiesList` when the component is
created.

We also prevent click events from passing through our `BasePracticeIcon`
and `BaseFlyout` by adding the `.stop` event modifier to them click event
handlers.

Lastly, we change the `overflow` style of our flyouts so scroll isn't
automatically enabled.