Prevent activity number generation
This commit stops our activity numbers from AUTO-GENERATING.
I wasn't sure if the number should be a plain number or if it could
have dashes etc. so I didn't put any validation on the input.

This commit also labels the activity view's route as activity/id/2, for
example. This gives meaning to the "weird" numbers.

If this doesn't look good, we could change the route to take
`activityNumber` as a its param instead. Then we would use that to
retrieve the activity. For example, activity/51896. I'm not sure what
lengths the numbers are going to be, or if they have dashes in them,
 so let me know what you think we should do.