Adds dynamic label classNames
This commit uses an onDraw event listener to dynamically decide
classnames for inner chart labels. The `createLabel` method, which took
11 arguments with little documentation on how to set the values, was
very hard to work with. This was a reasonable alternative to that.

The event also conditionally decides whether to show the label depending
on its value. Currently, we don't show the label if the value is 0, but
we can refactor this to use a reduce method and decide based off the
percentage.

The `showLabel` Chartist option doesn't seem to work, so we work around
that by setting the label's `display` in the className we add.

Lastly, we separate the data and options parameters we feed to the Pie
chart so our code is more readable.