Add budget scales
This commit adds vSelect component to our budget input. The user can
choose between the following options: "Thousand", "Million", and
"Billion".

The budget is automatically converted to a total when we
update the activity in our store. In addition, the app automatically
detects which scale the budget is in when pulling from the store. It
defaults to 1000.