Refactor Chartist.Pie implementation
This commit makes some small adjustments to how we are rendering the pie
chart. First we are `labelInterpolationFnc` option callback to
manipulate the label string - this will add the currency symbol to
charts that have the `isCurrency` prop set, as well as reduce the string
down to its base with a metric suffix (10,000 becomes 10k).

Second, this commit refactors how we are adding css classes to the
labels `text` elements. We use the on draw event, but instead match the
`context.index` to the `seriesData` index to match the css classnames.
We then append the css class to the `context.element`s classList.

Lastly, this commit makes some tweaks to the design of the charts,
specifically rendering them as donuts instead of pies.