Set `id` param for `/activity` route to be optional
This commit configures the `id` param to be option when passed to the
`/activity` route. This will set the `/activity` route to be a create
route, and the `activity/{id}` route to be the edit route. This commit
also decouples the `activityId` and `activityNumber` - the
`activityNumber` should be any string that represents the domain id of
the activity, while `activityId` should represent the activity entity
id within the Vue application's data store.