Saves activity on change
This commit saves activities when their value changes and that input is
blurred. Prior to this, the input was saved every time the input was
blurred.

The youth centric and activity type are still handled using `@input`.
When using `@change` on these components, the values would not
change inside of `ActivityInput`. Due to how those components work,
their new value will never be the same as their old value, so `@input` is
effectively the same as `@change`.