Clear console error caused by NavFooter
This commit fixes an issue caused by calling `this.$t` in the `NavFooter`
component. `this.$t` is a method added to the Vue prototype by the i18n
module. The issue is that `this.$t` tries to access the `this.i18n`
variable within the same prototype, but `this.i18n` is null (probably due
to a bug).

The work around for this ends up being to import the i18n module into
the component that is missing it and calling the `t()` method directly.