Accesses vue-i18n without local import
This commit accessed the `$t` method of the i18n module by calling the
`$root` method of the instance first. The `NavFooter` component is
deleted immediately after being rendered, which causes the `i18n` package
to set `_i18n` to `null`. `The `$root` method is still accessible and
must be used instead.