Refactor Results, Assessments page
This commit is a big overhaul of how tasks are displayed on the Results
and Assessments page. It allows the user to click on a task in list
view, and instead of it being expanded, it opens an ActivityTray
component on the right of the table with more details about the
activity. Here comments can be added, assessments can be made and
eventually you will be able to edit the activity inline.

This commit also changes the comments a bit, changing their relationship
to belong to activities directly. This is due to a poorly written task
requirement.