Refactor ActivityRecommendationInput
This commit generalizes `ActivityRecommendationInput` and renames it to
`ActivityiesItemInput`.  The `recommendationType` prop is separated into
the `action` and `inputType` props. This allows us to greatly simplify
the update and delete methods by reducing conditional statements and
using template strings instead.

The `globalRecommendation` Boolean prop is removed and combined with
the new `inputType` prop.

We also use the `&&` and spread operators to conditionally determine
which properties we include in the `data` object we send to the store.