Goes to new route when activity is submitted
Previously, when we were on `path/activity` and filled in a new activity,
the form would clear out and the activity would automatically be added
to the store. To prevent the form clearing, this commit changes the route
to the newly generated activity.

This commit also keeps track of the scroll position before the route
change and sets it to the same pos afterwards.