Fixes vue-select search bugs
This commit addresses issues created by the way `vue-select`'s search
funtionality was implemented. Two elements, a span `.vs__selected` and
an input `.vs__search` cause positioning bugs when the `vue-select` is
made too small.

When the dropbox is opened, the span is placed under the input and grayed
out, thus mimicing a placeholder. When the component is too small, there
is not enough room for this to occur. We can fix this by stopping the
span from being absolutely positioned and shrinking the input element.
The input element can't be removed completely or the component bugs even
more.

We do these only when the `searchable`  prop is turned off. When
`searchable` is turned on, the component is always large that this isn't
an issue. Since `BaseFormSelect` uses global styles, we only need to
implement the changes there.

This commit also removes the repetitive bottom border from the budget's
scale selector.