Fixes pie charts
This commit fixes issues caused by dividing by 0 when the budget or
total youth activities was 0. Normally, that gives us `NaN`, but here we
check before doing any calculations and give a value of 0 when that
happens. In the final file, the value will be 0 instead of `NaN`.

Now, in situations where everything is 0 (ex. no youth activites), we give
every category an equal piece of the pie (1) and simply change the labels
to "0". This change only occurs in the `Chart` component, so it doesn't
affect the final CSV file. It's just so the charts and legends aren't
empty.

This commit also defines a minimum percentage for which slices appear
on the chart, 3%. I doubled that to make a minimum percentage for labels
to appear. This makes the charts less crowded, removing tiny slices.

Lastly, we use `toFixed()` to show a maximum of 1 decimal place on the
charts. Now the labels aren't overflowing.