Navigate to completed activity
We want the app to auto save with each change to an activity, in case
the user is interrupted, however, we also want to give the user the
opportunity to review the activity after filling out the form. This
commit navigates the user to the completed activity once the activity
form is filled out, allowing them to review the form. They can then
click the Save and Add Another Activity button to add a new activity.
This button technically does not save the data, but reassures the user
that a save has happened.