Clean up activity tray
This commit cleans up the activity tray styles, and adds a button so
that we can edit the activity from the tray. It also updates the Go Back
button in the Activity component so that it returns to the previous page
in the history, allow us to go back to the analysis page after editing
an activity.