Reformats ActivityList on /summary
This commit adds an edit button, removes all EIP icons, and adds the
`activityNumber` to the `ActivityNumber` component in `/summary`.

This commit also increases the left box when using `activityNumber`s
(on `/summary`).

We also simplify the props we give to `ActivitiesItemAssessment`. Now it
takes the whole activity object in one prop.