Simplify activity saving
This commit simplifies the activity saving process. When an activity is
input, the user will need to press the save button to submit a new
activity. This will clear the form and take the user to the top of the
form. If the users is editing an activity, the activity will be saved
with each input change. In order to avoid confusion, the save activity
button is still present in the edit view. From the edit activity view,
the user also has the option to add another activity or save the current
activity.