Refactor NavBreadcrumbs to use global state
In order to incrementally indicate the progress of sub steps in the
NavBreadcrumbs component, this commit refactors the application to use
global state to set the progress of the user through the app. This
allows us to indicate easily how far they have progress through the app.
This commit also indicates which sub step the user is on and partially
fills the progress bar accordingly.