Refactor save
This commit refactors the activity saving workflow. This commit
introduces a global emitter bus that allows us to trigger an event in
one component, and listen to it in another component, regardless of
their relationship. This is used in the `NavFooter` to call an action
that triggers an event/method in the `ActivityInput` component.

This commit configures the `NavFooter` to save the Activity instead of
the `ActivityInput` component.