Update pdf timestamp
In order to overwrite the saved pdf, but still maintain a
timestamp/version of the saved pdf, this commit updates the timestamp
used. Instead of using a unix timestamp as the pdf suffix, this commit
uses the dd-mm-yyyy format as the suffix. This will allow users to
overwrite pdfs on the same day, while working, and also maintain a
version history when working over several days.