Refactors budget field
This commit adds a ones scale for the activity budget. The scale is
displayed as "--".

This commit also uses commas instead of periods in the french locale.
The decimal automatically changes everytime a user changes locals. We
use a custom validator for 'decimals' so we can use commas in the french
version.