BaseProgressBar shows accuracte percentage
This commit fixes the `BaseProgressBar` on the `/results` page. Normally,
we should be able to use `query().whereHas()` to retrieve activities
which at least one assessment with a value of `'yes'`. whereHas seems to be
glitched so we make an activities query and filters that to find how many.
`'yes'` assessments we have.

We also edit the `ActivitiesList` on `/results` to only show youth
activities since the header says as much.