Combines csv exports into a single file
This commit takes the 4 `.csv` files we originally exported and combines
them, separating the tables by 2 new lines. When importing, we parse the
file and `.split('\n\n')` to get the individual fields.

Since the store increments activities' ids when adding them, we have to
update the id of the imported activites. We also needed to update the
`activity_id` of the related assessments; otherwise, the assessments
wouldn't be linked to the right activity.