Fixes vue-select padding bug
Normally, `vue-select` rejects `mousedown` events when they occur on the
`padding` of an element. This commit creates a new event listener that
toggles the dropdown when a user clicks on a `padding`.

We use `jquery` to work with `vue-select`'s elements and know whether or
not we are actually clicking` padding`.