Fixes activity saving bug
This commit checks if the input fields are filled when creating a new
activity. We do this because the save buttons in the `NavFooter` always
run `addActivity()` via the `$eventHub` and end up saving blank
information.

This commit also fixes an issue with `FileExport`. Since default props
run the `this.$t` function and save the value, there was no way of
updating the text when the locale changed. This commit extracts the
default prop and calls it dynamically in the template.