Fix pdf download
This commit fixes the pdf export, so that it includes the default file
extension in the same manner as the json/csv exports. This commit also
fixes an issue where the Learn More About EIPs button in the activity
input form was opening into a blank new window in the Electron build.