Update population prospects data and remove activity budget scale
This commit updates the population prospect data source url to reflect
the new 2019 data. While the data itself was up to date, all of the
sources citing the data linked to the incorrect spreadsheet.

This commit also removes the activity budget scale field, as it was
causing a good deal of confusion. Instead, we are using a locale
specific budget field for formatting budget inputs using the correct
locale.