Fix checkbox issue
This commit fixes an issue with the checkbox for the youth centric
definition. Essentially, the checkbox was not bound correctly to the youth
centric value, and while the bug was over looked in Chrome, Edge did not
like it. This prevented any activities from being marked as youth centric,
and thus showing up in the activity tray.